% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uy_departamentos}
\alias{uy_departamentos}
\title{Departamentos de Uruguay}
\format{
Un objeto \code{sf} con 19 filas y 4 columnas:
\describe{
\item{NOMDEPTO}{Nombre del departamento en mayúsculas}
\item{CDEPTO_ISO}{Código ISO del departamento}
\item{DEPTO}{Código numérico del departamento}
\item{geom}{Geometría del polígono (MULTIPOLYGON), CRS: WGS 84 (EPSG 4326)}
}
}
\source{
Instituto Nacional de Estadística (INE), Censo 2023.
\url{https://www.gub.uy/instituto-nacional-estadistica/cartografia}
}
\usage{
uy_departamentos
}
\description{
Límites de los 19 departamentos de la República Oriental del Uruguay
según el Censo 2023 del Instituto Nacional de Estadística (INE).
}
\note{
Los datos tienen topología simplificada y se consideran referenciales.
Para uso geodésico oficial consultar el INE Uruguay.
}
\keyword{datasets}
